Removes harmful check that prohibits startup
The check that has been removed is not useful, because the file size
only tells us something if the new file could not have grown beyond the
size of the old file. This is dependant on the growth rate of the file
and the amount of time elapsed, which we can only make assumptions
about.